refactor: extract foreign-chain-health-check library from config-tester
Move the golden reference transactions, per-provider checks, and the
check_all_providers orchestration out of the foreign-chain-config-tester
CLI into a new production library crate, so the MPC node can reuse them
for a startup RPC health check without depending on the operator CLI
(which carries clap/serde_yaml/toml).

No behavior change: the CLI keeps its arg parsing, config-shape
detection, and table rendering, and now consumes the shared crate.

Part of #3764.
