+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
     "crates/devnet",
     "crates/e2e-tests",
     "crates/foreign-chain-config-tester",
+    "crates/foreign-chain-health-check",
     "crates/foreign-chain-inspector",
     "crates/foreign-chain-rpc-auth",
     "crates/foreign-chain-rpc-interfaces",
@@ -54,6 +55,7 @@ attestation = { path = "crates/attestation" }
 chain-gateway = { path = "crates/chain-gateway" }
 chain-gateway-test-contract = { path = "crates/chain-gateway-test-contract" }
 contract-history = { path = "crates/contract-history" }
+foreign-chain-health-check = { path = "crates/foreign-chain-health-check" }
 foreign-chain-inspector = { path = "crates/foreign-chain-inspector" }
 foreign-chain-rpc-auth = { path = "crates/foreign-chain-rpc-auth" }
 foreign-chain-rpc-interfaces = { path = "crates/foreign-chain-rpc-interfaces" }
