+++ crates/foreign-chain-config-tester/Cargo.toml
@@ -10,24 +10,13 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = { workspace = true }
-bs58 = { workspace = true }
 clap = { workspace = true }
-foreign-chain-inspector = { workspace = true }
-foreign-chain-rpc-auth = { workspace = true }
-foreign-chain-rpc-interfaces = { workspace = true }
-hex = { workspace = true }
-http = { workspace = true }
+foreign-chain-health-check = { workspace = true }
 mpc-node-config = { workspace = true }
 serde = { workspace = true }
 serde_yaml = { workspace = true }
 tokio = { workspace = true }
 toml = { workspace = true }
-
-[dev-dependencies]
-assert_matches = { workspace = true }
-httpmock = { workspace = true }
-near-mpc-bounded-collections = { workspace = true }
-serde_json = { workspace = true }
 
 [lints]
 workspace = true
