+++ crates/foreign-chain-health-check/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "foreign-chain-health-check"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+bs58 = { workspace = true }
+foreign-chain-inspector = { workspace = true }
+foreign-chain-rpc-auth = { workspace = true }
+foreign-chain-rpc-interfaces = { workspace = true }
+hex = { workspace = true }
+http = { workspace = true }
+mpc-node-config = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+httpmock = { workspace = true }
+near-mpc-bounded-collections = { workspace = true }
+serde_json = { workspace = true }
+
+[lints]
+workspace = true
